refactor(fn): reorganized cdk folder structure
