+++ libs/fn/src/lib/cdk/focusable-list/focusable-item.directive.ts
@@ -3,11 +3,11 @@ import { BooleanInput, coerceBooleanProperty } from '@angular/cdk/coercion';
 import { ReplaySubject } from 'rxjs';
 import { takeUntil } from 'rxjs/operators';
 import { FN_FOCUSABLE } from './focusable.tokens';
-import { FN_DISABLED } from '../tokens/disabled';
-import { FN_READONLY } from '../tokens/readonly';
+import { FN_DISABLED } from '../disabled/fn-disabled.token';
+import { FN_READONLY } from '../readonly/fn-readonly.token';
 import { BaseFocusableBehavior } from '../common-behaviors/base-focusable-behavior';
-import { DisabledBehavior } from '../interfaces/disabled-behavior.interface';
-import { ReadonlyBehavior } from '../interfaces/readonly-behavior.interface';
+import { DisabledBehavior } from '../disabled/disabled-behavior.interface';
+import { ReadonlyBehavior } from '../readonly/readonly-behavior.interface';
 import { DestroyedBehavior } from '../common-behaviors/destroyed-behavior';
 import { HasElementRef } from '../HasElementRef';
