+++ libs/fn/src/lib/cdk/index.ts
@@ -1,28 +1,16 @@
 export * from './cdk.module';
 
-export * from './disabled/disabled-behavior.directive';
-export * from './disabled/disabled-behavior.module';
-export * from './disabled/set-disabled-state';
+export * from './disabled';
 
-export * from './interfaces/disabled-behavior.interface';
-
-export * from './readonly/readonly-behavior.directive';
-export * from './interfaces/readonly-behavior.interface';
+export * from './readonly';
 
 export * from './directives/clicked/clicked.directive';
 
-export * from './selectable-list/selection.service';
-export * from './selectable-list/selectable-list.module';
-export * from './selectable-list/selectable-list.tokens';
-export * from './selectable-list/selectable-item.token';
-export * from './selectable-list/select-component-root.token';
+export * from './selectable-list';
 
 export * from './resize/resize.module';
 
-export * from './focusable-list/focusable-list.service';
-export * from './focusable-list/focusable-behavior.module';
-export * from './focusable-list/selectable-to-focusable-item';
-export * from './focusable-list/focusable.tokens';
+export * from './focusable-list';
 
 export * from './mixins/can-assign-additional-classes';
 export * from './mixins/has-tab-index';
@@ -31,8 +19,4 @@ export * from './mixins/can-be-disabled';
 export * from './common-behaviors/base-focusable-behavior';
 export * from './common-behaviors/destroyed-behavior';
 
-export * from './tokens/disabled';
-export * from './tokens/readonly';
-
 export * from './HasElementRef';
-export { ReadonlyBehaviorModule } from './readonly/readonly-behavior.module';
